requirements.txt
tldextract
dnspython

